refactor(panel): isolate restart and terminal restoration
